# api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Prompt Injection Defense API
  description: |
    ## Real-time Prompt Injection Detection API
    
    **Enterprise-grade security for AI applications** - Detect and prevent prompt injection attacks with <50ms latency using advanced AI models.
    
    ### Key Features
    
    🛡️ **Multi-Model Detection**
    - ProtectAI DeBERTa v3 (specialized prompt injection classifier)
    - Meta Llama Prompt Guard 2-86M (lightweight jailbreak detection) 
    - Google Gemini 2.0 Flash (advanced semantic analysis)
    
    ⚡ **High Performance**
    - Sub-50ms detection latency
    - Concurrent processing with fallback mechanisms
    - Automatic failover and retry logic
    
    🔐 **Enterprise Security**
    - API key authentication with rate limiting
    - Secure webhook delivery with signature verification
    - Comprehensive audit logging and analytics
    
    ### Threat Types Detected
    
    | Threat Type | Description |
    |-------------|-------------|
    | **Jailbreak** | Attempts to bypass AI safety guardrails |
    | **System Prompt Leak** | Attempts to extract internal AI instructions |
    | **Data Extraction** | Attempts to access training data or user info |
    | **Injection** | Malicious input designed to manipulate AI behavior |
    | **Encoding Attack** | Obfuscated attacks using Base64, hex, etc. |
    | **Delimiter Attack** | Using special characters to break context |
    
    ### Authentication
    
...
    
    ```bash
...
    
...
    ```
    
...
servers:
...
security:
  - APIKeyHeader: []
  - APIKeyBearer: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Create new API key
      description: |
        **Generate API Key** - Creates a new API key with specified rate limits and usage tracking.
        
        **Default Limits:**
        - 60 requests per minute
        - 10,000 requests per day
        
        **Custom Limits:**
        - Up to 1,000 requests per minute
        - Up to 100,000 requests per day
        
        **Important:** Store the returned API key securely as it won't be shown again.
      operationId: create_api_key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIKeyRequest'
            example:
              name: "Development API Key"
              rate_limit_per_minute: 100
              rate_limit_per_day: 50000
      responses:
        '200':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
              example:
                api_key: "pid_1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
                key_id: "key_abc123def456"
                name: "Development API Key"
                rate_limit_per_minute: 100
                rate_limit_per_day: 50000
                created_at: "2024-01-15T10:30:00Z"
                expires_at: null
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get API key usage statistics
      description: |
        **Profile Information** - Returns current usage statistics, rate limits, and activity history for the authenticated API key.
      operationId: get_api_key_profile
      security:
        - APIKeyHeader: []
        - APIKeyBearer: []
      responses:
        '200':
          description: Profile information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageStats'
        '401':
          $ref: '#/components/responses/AuthenticationError'
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication via X-API-Key header
    APIKeyBearer:
      type: http
      scheme: bearer
      description: API Key authentication via Authorization Bearer header
  schemas:
...
    APIKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Human-readable name for the API key
        rate_limit_per_minute:
          type: integer
          minimum: 1
          maximum: 1000
          description: Custom rate limit per minute
        rate_limit_per_day:
          type: integer
          minimum: 1
          maximum: 100000
          description: Custom rate limit per day
    APIKeyResponse:
      type: object
      properties:
        api_key:
          type: string
          description: The generated API key (store securely)
          example: "pid_1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
        key_id:
          type: string
          description: Unique identifier for this API key
        name:
          type: string
          description: Human-readable name
        rate_limit_per_minute:
          type: integer
          description: Requests per minute allowed
        rate_limit_per_day:
          type: integer
          description: Requests per day allowed
        created_at:
          type: string
          format: date-time
          description: When the key was created
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: When the key expires
    UsageStats:
      type: object
      properties:
        key_id:
          type: string
          description: API key identifier
        name:
          type: string
          description: API key name
        requests_today:
          type: integer
          description: Requests made today
        requests_this_minute:
          type: integer
          description: Requests in current minute
        rate_limit_per_minute:
          type: integer
          description: Rate limit per minute
        rate_limit_per_day:
          type: integer
          description: Rate limit per day
        total_requests:
          type: integer
          description: Total requests ever
        malicious_detections:
          type: integer
          description: Total malicious content detected
        last_used_at:
          type: string
          format: date-time
          nullable: true
          description: Last usage timestamp
...
  responses:
    ValidationError:
      description: Request validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
          example:
            detail:
              - loc: ["body", "text"]
                msg: "field required"
                type: "value_error.missing"
    AuthenticationError:
      description: Authentication error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
          example:
            detail:
              error: "API key required"
              message: "Provide API key in X-API-Key header or Authorization: Bearer header"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
          example:
            detail: "Detection service unavailable"